Corrected Bottom Loss/Too Shallow
The rayleigh reflection model was previous taking the grazing angle to the normal, hence the complement. Adjusted the angle inside of rayleigh to now take the grazing angle from the horizontal.

Too shallow logic was being used at the beginning of a reflection but was later being ignored when it was required to actually perform the reflection. Added another instance of too shallow to rectify this issue.

Updated format of reflect_loss_netcdf and its files. Test updated to allow for correct format.